LayoutTests/fast/js/dfg-dead-min-one-arg-expected.txt

Tests that a dummy use of Math.min(a) at least speculates that its argument is indeed a number.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS result is 42
PASS ok is 150
PASS result is 37
PASS ok is 151
PASS result is 37
PASS ok is 152
PASS result is 37
PASS ok is 153
PASS result is 37
PASS ok is 154
PASS result is 37
PASS ok is 155
PASS result is 37
PASS ok is 156
PASS result is 37
PASS ok is 157
PASS result is 37
PASS ok is 158
PASS result is 37
PASS ok is 159
PASS result is 37
PASS ok is 160
PASS result is 37
PASS ok is 161
PASS result is 37
PASS ok is 162
PASS result is 37
PASS ok is 163
PASS result is 37
PASS ok is 164
PASS result is 37
PASS ok is 165
PASS result is 37
PASS ok is 166
PASS result is 37
PASS ok is 167
PASS result is 37
PASS ok is 168
PASS result is 37
PASS ok is 169
PASS result is 37
PASS ok is 170
PASS result is 37
PASS ok is 171
PASS result is 37
PASS ok is 172
PASS result is 37
PASS ok is 173
PASS result is 37
PASS ok is 174
PASS result is 37
PASS ok is 175
PASS result is 37
PASS ok is 176
PASS result is 37
PASS ok is 177
PASS result is 37
PASS ok is 178
PASS result is 37
PASS ok is 179
PASS result is 37
PASS ok is 180
PASS result is 37
PASS ok is 181
PASS result is 37
PASS ok is 182
PASS result is 37
PASS ok is 183
PASS result is 37
PASS ok is 184
PASS result is 37
PASS ok is 185
PASS result is 37
PASS ok is 186
PASS result is 37
PASS ok is 187
PASS result is 37
PASS ok is 188
PASS result is 37
PASS ok is 189
PASS result is 37
PASS ok is 190
PASS result is 37
PASS ok is 191
PASS result is 37
PASS ok is 192
PASS result is 37
PASS ok is 193
PASS result is 37
PASS ok is 194
PASS result is 37
PASS ok is 195
PASS result is 37
PASS ok is 196
PASS result is 37
PASS ok is 197
PASS result is 37
PASS ok is 198
PASS result is 37
PASS ok is 199
PASS result is 37
PASS successfullyParsed is true

TEST COMPLETE
